🔢 refactor(quantization): switch storage format from int8 to fp8 (e4m3)
Matches the reference implementation in ~/experiment (partial_moe.py:FP8Linear,
17c878d) so loss measurements transfer between the two codebases. Storage
format and forward math only — scope, attachment points and the
dequantization-transparent state_dict() contract are unchanged.

  float8_e4m3fn / QMAX 448, cast without an explicit round or clamp.
  Forward dequantizes in the activation dtype, matching the reference exactly
  (including the stacked-experts path); serialization keeps fp32 arithmetic.
  model.quantization: "int8" -> "fp8"; CONNITO_ALLOW_VALIDATOR_INT8 ->
  CONNITO_ALLOW_VALIDATOR_FP8.

Both formats are one byte per weight, so this buys no memory (measured
identical at 8.428 GB) and costs 3.19x fidelity: 2.645% relative weight error
against int8's 0.829%. Comparability is the whole return.

Three traps fp8 has that int8 did not, each found by probing on an L40S and
each now pinned by a test:

  * float8_e4m3fn IS a floating dtype, so Module.to(dtype=) upcasts the weight
    buffer and the memory saving evaporates with nothing raised. Both _apply
    overrides re-assert the storage dtype. Casting back is exactly lossless.
  * torch.isfinite raises NotImplementedError for fp8 on CPU and CUDA alike.
    The miner's post-setup guard filters buffers on is_floating_point(), which
    int8 failed and fp8 passes, so this would have crashed every fp8 miner at
    startup. quantization.all_finite uses torch.isnan, which is exhaustive
    here: e4m3fn has no infinity encoding.
  * e4m3fn overflow produces NaN, not saturation. Safe only because the per-row
    scale caps |w/scale| at 448.000031 and everything up to 464 rounds back to
    448 — verified over a 200k-row adversarial sweep.

Two dequantization functions, deliberately: in fp16 a row whose scale rounds to
zero (amax < 1.3351e-5) would be silently deleted from a checkpoint by the
activation-dtype multiply, so state_dict() keeps fp32 arithmetic.

The validator refusal now separates int8-measured evidence from fp8 inference:
the rank gate was only ever run under int8 and int8 failed it; fp8 is coarser
and has not been gated in its own right. A test pins that wording.

Also: FP8Linear._save_to_state_dict moves buffers to CPU before dequantizing
rather than after, which is what its comment already claimed (save-path VRAM
overhead 0.168 GB -> 0.000 GB); and _load_from_state_dict no longer allocates a
second full stacked clone per graft just to populate a debug log field.

+++ connito/shared/helper.py
@@ -42,7 +42,7 @@ def infer_storage_dtype(model: "torch.nn.Module", default: torch.dtype = torch.f
     """Best-effort read of the half precision a built model stores weights at.
 
     Order-independent, unlike `next(model.parameters()).dtype`: it skips fp32
-    (the miner upcasts *trainable* params to fp32 for AdamW) and int8 (runtime
+    (the miner upcasts *trainable* params to fp32 for AdamW) and fp8 (runtime
     quantization moves converted weights into buffers), so it answers with the
     model's actual storage precision wherever that tensor happens to sit.
     """
